Add a way to view the table at the end of the game
Resolves:
https://github.com/joffrey-bion/seven-wonders/issues/59

+++ sw-ui/src/jsMain/kotlin/org/luxons/sevenwonders/ui/components/game/Board.kt
@@ -115,7 +115,7 @@ private fun ChildrenBuilder.wonderComponent(wonder: ApiWonder, military: Militar
                 maxWidth = 95.pct // same as wonder
 
                 // bring to the foreground on hover
-                hover { zIndex = integer(1000) }
+                hover { zIndex = integer(50) }
             }
             img {
                 src =  "/images/wonders/${wonder.image}"
